refactor(launch): split LaunchPage into pages/launch/* submodules
